// src/payroll/interfaces/payroll-calculation.interface.ts
export interface PayrollCalculationData {
  id?: string;
  sr?: number;
  name?: string;
  designation?: string;
  department?: string;
  month?: string;
  year?: number;
  totalDays: number;
  offDays?: number;
  leaveTaken?: number;
  workedDays: number;
  ctc: number;
  dailyRate?: number;
  hourlyRate?: number;
  offDaysWorked?: number;
  offDayAmount?: number;
  holidayWorked?: number;
  holidayAmount?: number;
  leaveSalary?: number;
  cashAdvance?: number;
  penaltyPoints?: number;
  total?: number;
  visaCost?: number;
  absences?: number;
  unauthorizedAbsences?: number;
  lateHours?: number;
  authAbsenceDeduction?: number;
  unauthAbsenceDeduction?: number;
  tardiness?: number;
  fines?: number;
  cleaningFees?: number;
  allDeductions?: number;
  overtimeHours?: number;
  overtimeAmount?: number;
  netDeductions?: number;
  extraFromManager?: number;
  januaryNetSalary?: number;
  targetRate?: number;
  backPayment?: number;
  totalJanuarySalary?: number;
  finalModification?: number;
  hrNotes?: string;
  beforeOT?: number;
  ot?: number;
  totalCalculated?: number;
  dfrnce?: number;
  deductions?: number;
  inDays?: number;
  isCalculated?: boolean;
}